requirements.txt
torch>=2.0
torchaudio>=2.0
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.65
umap-learn>=0.5.4
hdbscan>=0.8.33
soundfile>=0.12
librosa>=0.10
